RavenDB-11310 - Make some fields to private in Leader and Follower
